% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loadAllChunksCache.R
\name{loadAllChunksCache}
\alias{loadAllChunksCache}
\title{loadAllChunksCache}
\usage{
loadAllChunksCache(fileContainingChunks, pathForCache = paste0(sub(".Rmd", "",
  fileContainingChunks), "_cache/", names(which(sapply(c("html", "pdf", "word"),
  function(x) dir.exists(paste0(sub(".Rmd", "", fileContainingChunks),
  "_cache/", x, "/")))))[1], "/"))
}
\arguments{
\item{fileContainingChunks}{The path to your Rmd file}

\item{pathForCache}{The path to the cache if not by default}
}
\value{
messages with the name of the loaded chunks
}
\description{
loadAllChunksCache
}
\examples{
\dontrun{
loadAllChunksCache("input.Rmd")
}
}
